refactor WolfBiomeConfig
variants being a dedicated field in the WolfBiomeConfig, it's resourceLocation is
no longer relied on to figure out which variant is bound to the config.
Also variant field being a list so that mutiple variant can bound to a biome
or mutiple biome interchangebly
Refactor Util where WolfBiomeConfig is queried based on biome. Seperating the logic into seperate function,
also moves getExtraSpawnable to WolfvariantUtil instead.